+++ backend/Cargo.toml
@@ -42,6 +42,7 @@ tower_governor = { version = "0.5.0", features = ["axum"] }
 # Redis caching for hot data (#714)
 redis = { version = "0.27", features = ["tokio-comp", "connection-manager"] }
 sysinfo = "0.29"
+predifi-errors = { path = "../contract/contracts/predifi-errors", features = ["std"] }
 
 
 [dev-dependencies]
