+++ contract/contracts/predifi-errors/Cargo.toml
@@ -9,3 +9,6 @@ crate-type = ["rlib"]
 
 [dependencies]
 soroban-sdk = { workspace = true }
+
+[features]
+std = []
